+++ src/core/lombok/eclipse/handlers/EclipseHandlerUtil.java
@@ -48,6 +48,7 @@ import lombok.core.TypeResolver;
 import lombok.core.configuration.NullCheckExceptionType;
 import lombok.core.debug.ProblemReporter;
 import lombok.core.handlers.HandlerUtil;
+import lombok.eclipse.Eclipse;
 import lombok.eclipse.EclipseAST;
 import lombok.eclipse.EclipseNode;
 import lombok.experimental.Accessors;
@@ -411,19 +412,17 @@ public class EclipseHandlerUtil {
 	}
 	
 	public static Annotation[] copyAnnotations(ASTNode source, Annotation[]... allAnnotations) {
-		boolean allNull = true;
-		
-		List<Annotation> result = new ArrayList<Annotation>();
+		List<Annotation> result = null;
 		for (Annotation[] annotations : allAnnotations) {
 			if (annotations != null) {
-				allNull = false;
 				for (Annotation annotation : annotations) {
+					if (result == null) result = new ArrayList<Annotation>();
 					result.add(copyAnnotation(annotation, source));
 				}
 			}
 		}
-		if (allNull) return null;
-		return result.toArray(new Annotation[0]);
+		
+		return result == null ? null : result.toArray(new Annotation[0]);
 	}
 	
 	public static boolean hasAnnotation(Class<? extends java.lang.annotation.Annotation> type, EclipseNode node) {
@@ -1215,8 +1214,9 @@ public class EclipseHandlerUtil {
 	 * Inserts a field into an existing type. The type must represent a {@code TypeDeclaration}.
 	 * The field carries the &#64;{@link SuppressWarnings}("all") annotation.
 	 */
-	public static EclipseNode injectFieldSuppressWarnings(EclipseNode type, FieldDeclaration field) {
-		field.annotations = createSuppressWarningsAll(field, field.annotations);
+	public static EclipseNode injectFieldAndMarkGenerated(EclipseNode type, FieldDeclaration field) {
+		field.annotations = addSuppressWarningsAll(type, field, field.annotations);
+		field.annotations = addGenerated(type, field, field.annotations);
 		return injectField(type, field);
 	}
 	
@@ -1261,7 +1261,8 @@ public class EclipseHandlerUtil {
 	 * Inserts a method into an existing type. The type must represent a {@code TypeDeclaration}.
 	 */
 	public static EclipseNode injectMethod(EclipseNode type, AbstractMethodDeclaration method) {
-		method.annotations = createSuppressWarningsAll(method, method.annotations);
+		method.annotations = addSuppressWarningsAll(type, method, method.annotations);
+		method.annotations = addGenerated(type, method, method.annotations);
 		TypeDeclaration parent = (TypeDeclaration) type.get();
 		
 		if (parent.methods == null) {
@@ -1303,9 +1304,10 @@ public class EclipseHandlerUtil {
 	 * @param type New type (class, interface, etc) to inject.
 	 */
 	public static EclipseNode injectType(final EclipseNode typeNode, final TypeDeclaration type) {
-		type.annotations = createSuppressWarningsAll(type, type.annotations);
+		type.annotations = addSuppressWarningsAll(typeNode, type, type.annotations);
+		type.annotations = addGenerated(typeNode, type, type.annotations);
 		TypeDeclaration parent = (TypeDeclaration) typeNode.get();
-
+		
 		if (parent.memberTypes == null) {
 			parent.memberTypes = new TypeDeclaration[] { type };
 		} else {
@@ -1319,8 +1321,31 @@ public class EclipseHandlerUtil {
 	}
 	
 	private static final char[] ALL = "all".toCharArray();
+	private static final char[] JUSTIFICATION = "justification".toCharArray();
+	private static final char[] GENERATED_CODE = "generated code".toCharArray();
+	private static final char[] LOMBOK = "lombok".toCharArray();
+	private static final char[][] JAVAX_ANNOTATION_GENERATED = Eclipse.fromQualifiedName("javax.annotation.Generated");
+	private static final char[][] EDU_UMD_CS_FINDBUGS_ANNOTATIONS_SUPPRESSFBWARNINGS = Eclipse.fromQualifiedName("edu.umd.cs.findbugs.annotations.SuppressFBWarnings");
 	
-	public static Annotation[] createSuppressWarningsAll(ASTNode source, Annotation[] originalAnnotationArray) {
+	public static Annotation[] addSuppressWarningsAll(EclipseNode node, ASTNode source, Annotation[] originalAnnotationArray) {
+		Annotation[] anns = addAnnotation(source, originalAnnotationArray, TypeConstants.JAVA_LANG_SUPPRESSWARNINGS, new StringLiteral(ALL, 0, 0, 0));
+		
+		if (Boolean.TRUE.equals(node.getAst().readConfiguration(ConfigurationKeys.ADD_FINDBUGS_SUPPRESSWARNINGS_ANNOTATIONS))) {
+			MemberValuePair mvp = new MemberValuePair(JUSTIFICATION, 0, 0, new StringLiteral(GENERATED_CODE, 0, 0, 0));
+			anns = addAnnotation(source, anns, EDU_UMD_CS_FINDBUGS_ANNOTATIONS_SUPPRESSFBWARNINGS, mvp);
+		}
+		
+		return anns;
+	}
+	
+	public static Annotation[] addGenerated(EclipseNode node, ASTNode source, Annotation[] originalAnnotationArray) {
+		if (Boolean.FALSE.equals(node.getAst().readConfiguration(ConfigurationKeys.ADD_GENERATED_ANNOTATIONS))) return originalAnnotationArray;
+		return addAnnotation(source, originalAnnotationArray, JAVAX_ANNOTATION_GENERATED, new StringLiteral(LOMBOK, 0, 0, 0));
+	}
+	
+	private static Annotation[] addAnnotation(ASTNode source, Annotation[] originalAnnotationArray, char[][] annotationTypeFqn, ASTNode arg) {
+		char[] simpleName = annotationTypeFqn[annotationTypeFqn.length - 1];
+		
 		if (originalAnnotationArray != null) for (Annotation ann : originalAnnotationArray) {
 			char[] lastToken = null;
 			
@@ -1331,20 +1356,41 @@ public class EclipseHandlerUtil {
 				lastToken = ((SingleTypeReference) ann.type).token;
 			}
 			
-			if (lastToken != null && new String(lastToken).equals("SuppressWarnings")) return originalAnnotationArray;
+			if (lastToken != null && Arrays.equals(simpleName, lastToken)) return originalAnnotationArray;
 		}
 		
 		int pS = source.sourceStart, pE = source.sourceEnd;
 		long p = (long)pS << 32 | pE;
-		long[] poss = new long[3];
+		long[] poss = new long[annotationTypeFqn.length];
 		Arrays.fill(poss, p);
-		QualifiedTypeReference suppressWarningsType = new QualifiedTypeReference(TypeConstants.JAVA_LANG_SUPPRESSWARNINGS, poss);
-		setGeneratedBy(suppressWarningsType, source);
-		SingleMemberAnnotation ann = new SingleMemberAnnotation(suppressWarningsType, pS);
-		ann.declarationSourceEnd = pE;
-		ann.memberValue = new StringLiteral(ALL, pS, pE, 0);
+		QualifiedTypeReference qualifiedType = new QualifiedTypeReference(annotationTypeFqn, poss);
+		setGeneratedBy(qualifiedType, source);
+		Annotation ann;
+		if (arg instanceof Expression) {
+			SingleMemberAnnotation sma = new SingleMemberAnnotation(qualifiedType, pS);
+			sma.declarationSourceEnd = pE;
+			arg.sourceStart = pS;
+			arg.sourceEnd = pE;
+			sma.memberValue = (Expression) arg;
+			setGeneratedBy(sma.memberValue, source);
+			ann = sma;
+		} else if (arg instanceof MemberValuePair) {
+			NormalAnnotation na = new NormalAnnotation(qualifiedType, pS);
+			na.declarationSourceEnd = pE;
+			arg.sourceStart = pS;
+			arg.sourceEnd = pE;
+			na.memberValuePairs = new MemberValuePair[] {(MemberValuePair) arg};
+			setGeneratedBy(na.memberValuePairs[0], source);
+			setGeneratedBy(na.memberValuePairs[0].value, source);
+			na.memberValuePairs[0].value.sourceStart = pS;
+			na.memberValuePairs[0].value.sourceEnd = pE;
+			ann = na;
+		} else {
+			MarkerAnnotation ma = new MarkerAnnotation(qualifiedType, pS);
+			ma.declarationSourceEnd = pE;
+			ann = ma;
+		}
 		setGeneratedBy(ann, source);
-		setGeneratedBy(ann.memberValue, source);
 		if (originalAnnotationArray == null) return new Annotation[] { ann };
 		Annotation[] newAnnotationArray = new Annotation[originalAnnotationArray.length + 1];
 		System.arraycopy(originalAnnotationArray, 0, newAnnotationArray, 0, originalAnnotationArray.length);